Move PackageReference to PackageModel
